#1: Index distribution Added new methods for encoding/decoding operation responses.

+++ codec/src/main/java/ch/ethz/globis/disindex/codec/OpStatus.java
@@ -2,7 +2,7 @@ package ch.ethz.globis.disindex.codec;
 
 public class OpStatus {
 
-    public static final int SUCCESS = 0;
+    public static final byte SUCCESS = 0;
 
-    public static final int FAILURE = -1;
+    public static final byte FAILURE = -1;
 }
